% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acm0002_ontology.R
\name{read_acm0002_grid_shapes}
\alias{read_acm0002_grid_shapes}
\title{Load ACM0002 grid connection applicability shape}
\usage{
read_acm0002_grid_shapes()
}
\value{
A \code{sh_shape_graph} as returned by \code{\link[shapeR:read_shacl]{shapeR::read_shacl()}}.
}
\description{
Load ACM0002 grid connection applicability shape
}
